requirements.txt
biopython>=1.79
requests>=2.28.0
colorama>=0.4.4
pandas>=1.5.0
joblib>=1.2.0
tqdm>=4.64.0
numpy>=1.21.0
urllib3>=1.26.0
